+++ tests/test_modbus_helpers.py
@@ -186,127 +186,6 @@ async def test_async_close_client_with_async_close():
 # ---------------------------------------------------------------------------
 
 
-def test_get_signature_returns_none_for_non_inspectable():
-    """_get_signature returns None for objects that raise TypeError."""
-    from custom_components.thessla_green_modbus.modbus_helpers import _get_signature
-
-    # Built-in C functions raise TypeError on inspect.signature
-    result = _get_signature(len)
-    # May return a signature or None depending on Python version;
-    # the key thing is it doesn't raise
-    assert result is None or result is not None
-
-
-def test_get_signature_caches_result():
-    """_get_signature returns cached result on repeated calls."""
-    from custom_components.thessla_green_modbus.modbus_helpers import _get_signature
-
-    async def my_func(a, b): ...
-
-    r1 = _get_signature(my_func)
-    r2 = _get_signature(my_func)
-    assert r1 is r2
-
-
-# ---------------------------------------------------------------------------
-# _mask_frame — short hex edge case (line 124)
-# ---------------------------------------------------------------------------
-
-
-def test_mask_frame_empty():
-    from custom_components.thessla_green_modbus.modbus_helpers import _mask_frame
-
-    assert _mask_frame(b"") == ""
-
-
-def test_mask_frame_single_byte():
-    """Single byte → hex is 2 chars → mask covers them both."""
-    from custom_components.thessla_green_modbus.modbus_helpers import _mask_frame
-
-    result = _mask_frame(bytes([0x01]))
-    assert result.startswith("**")
-
-
-def test_mask_frame_multi_bytes():
-    from custom_components.thessla_green_modbus.modbus_helpers import _mask_frame
-
-    result = _mask_frame(bytes([0x01, 0x04, 0x00, 0x64]))
-    assert result.startswith("**")
-    assert "0064" in result
-
-
-# ---------------------------------------------------------------------------
-# _build_request_frame — uncovered func names (lines 147-183)
-# ---------------------------------------------------------------------------
-
-
-def test_build_request_frame_read_coils():
-    from custom_components.thessla_green_modbus.modbus_helpers import _build_request_frame
-
-    frame = _build_request_frame("read_coils", 1, [100], {"count": 8})
-    assert frame[0] == 1  # slave_id
-    assert frame[1] == 1  # function code for read_coils
-    assert len(frame) == 6
-
-
-def test_build_request_frame_read_discrete_inputs():
-    from custom_components.thessla_green_modbus.modbus_helpers import _build_request_frame
-
-    frame = _build_request_frame("read_discrete_inputs", 1, [200], {"count": 4})
-    assert frame[1] == 2  # function code
-
-
-def test_build_request_frame_write_register():
-    from custom_components.thessla_green_modbus.modbus_helpers import _build_request_frame
-
-    frame = _build_request_frame("write_register", 1, [100], {"value": 42})
-    assert frame[1] == 6  # function code for write single
-    assert len(frame) == 6
-
-
-def test_build_request_frame_write_registers():
-    from custom_components.thessla_green_modbus.modbus_helpers import _build_request_frame
-
-    frame = _build_request_frame("write_registers", 1, [100], {"values": [10, 20]})
-    assert frame[1] == 16  # function code 0x10
-    assert len(frame) == 7 + 4  # header(7) + 2 values * 2 bytes
-
-
-def test_build_request_frame_write_coil_true():
-    from custom_components.thessla_green_modbus.modbus_helpers import _build_request_frame
-
-    frame = _build_request_frame("write_coil", 1, [50], {"value": True})
-    assert frame[1] == 5  # function code for write coil
-    assert frame[4] == 0xFF  # 65280 = 0xFF00 → high byte
-
-
-def test_build_request_frame_write_coil_false():
-    from custom_components.thessla_green_modbus.modbus_helpers import _build_request_frame
-
-    frame = _build_request_frame("write_coil", 1, [50], {"value": False})
-    assert frame[4] == 0x00  # value 0 → high byte = 0
-
-
-def test_build_request_frame_unknown_func_returns_empty():
-    from custom_components.thessla_green_modbus.modbus_helpers import _build_request_frame
-
-    frame = _build_request_frame("unknown_function", 1, [], {})
-    assert frame == b""
-
-
-def test_build_request_frame_value_error_returns_empty():
-    from custom_components.thessla_green_modbus.modbus_helpers import _build_request_frame
-
-    # Pass non-integer address to trigger ValueError
-    frame = _build_request_frame("read_coils", 1, ["not_an_int"], {"count": 1})
-    assert frame == b""
-
-
-# ---------------------------------------------------------------------------
-# _calculate_backoff_delay — tuple jitter with inverted bounds (lines 212-214)
-# ---------------------------------------------------------------------------
-
-
 def test_calculate_backoff_jitter_tuple_inverted():
     """Jitter tuple with max < min is automatically swapped."""
     from custom_components.thessla_green_modbus.modbus_helpers import _calculate_backoff_delay
@@ -368,46 +247,6 @@ def test_calculate_backoff_zero_base():
 # ---------------------------------------------------------------------------
 
 
-def test_get_signature_typeerror_returns_none():
-    """inspect.signature raising TypeError returns None."""
-    import inspect
-    from unittest.mock import patch
-
-    from custom_components.thessla_green_modbus.modbus_helpers import _SIG_CACHE, _get_signature
-
-    def func():
-        pass
-
-    # Ensure not in cache
-    _SIG_CACHE.pop(func, None)
-
-    with patch.object(inspect, "signature", side_effect=TypeError("no sig")):
-        result = _get_signature(func)
-    assert result is None  # nosec B101
-
-
-def test_get_signature_valueerror_returns_none():
-    """inspect.signature raising ValueError returns None."""
-    import inspect
-    from unittest.mock import patch
-
-    from custom_components.thessla_green_modbus.modbus_helpers import _SIG_CACHE, _get_signature
-
-    def func():
-        pass
-
-    _SIG_CACHE.pop(func, None)
-
-    with patch.object(inspect, "signature", side_effect=ValueError("bad sig")):
-        result = _get_signature(func)
-    assert result is None  # nosec B101
-
-
-# ---------------------------------------------------------------------------
-# async_maybe_await_close — obj=None (line 96) and no close (line 100)
-# ---------------------------------------------------------------------------
-
-
 @pytest.mark.asyncio
 async def test_async_maybe_await_close_none():
     """obj=None returns immediately without error (line 96)."""
@@ -443,67 +282,6 @@ async def test_async_maybe_await_close_non_callable_close():
 # ---------------------------------------------------------------------------
 
 
-def test_chunk_register_range_zero_count():
-    """count<=0 returns empty list (line 418)."""
-    from custom_components.thessla_green_modbus.modbus_helpers import chunk_register_range
-
-    assert chunk_register_range(0, 0) == []  # nosec B101
-    assert chunk_register_range(5, -1) == []  # nosec B101
-
-
-def test_chunk_register_range_none_max_uses_const():
-    """max_block_size=None uses constant default (line 421)."""
-    from custom_components.thessla_green_modbus.modbus_helpers import chunk_register_range
-
-    chunks = chunk_register_range(100, 5, None)
-    assert len(chunks) == 1  # nosec B101
-    assert chunks[0] == (100, 5)  # nosec B101
-
-
-def test_chunk_register_range_max_zero_clamped_to_one():
-    """max_block_size<=0 is clamped to 1 (line 424)."""
-    from custom_components.thessla_green_modbus.modbus_helpers import chunk_register_range
-
-    chunks = chunk_register_range(0, 3, 0)
-    assert len(chunks) == 3  # nosec B101
-    assert all(c[1] == 1 for c in chunks)  # nosec B101
-
-
-# ---------------------------------------------------------------------------
-# chunk_register_values — empty, max=None, max<1 (lines 445, 448, 451)
-# ---------------------------------------------------------------------------
-
-
-def test_chunk_register_values_empty():
-    """Empty values list returns empty (line 445)."""
-    from custom_components.thessla_green_modbus.modbus_helpers import chunk_register_values
-
-    assert chunk_register_values(0, []) == []  # nosec B101
-
-
-def test_chunk_register_values_none_max_uses_const():
-    """max_block_size=None uses constant default (line 448)."""
-    from custom_components.thessla_green_modbus.modbus_helpers import chunk_register_values
-
-    chunks = chunk_register_values(10, [1, 2, 3], None)
-    assert chunks == [(10, [1, 2, 3])]  # nosec B101
-
-
-def test_chunk_register_values_max_zero_clamped_to_one():
-    """max_block_size<=0 is clamped to 1 (line 451)."""
-    from custom_components.thessla_green_modbus.modbus_helpers import chunk_register_values
-
-    chunks = chunk_register_values(0, [10, 20], 0)
-    assert len(chunks) == 2  # nosec B101
-    assert chunks[0] == (0, [10])  # nosec B101
-    assert chunks[1] == (1, [20])  # nosec B101
-
-
-# ---------------------------------------------------------------------------
-# _call_modbus — signature=None path, positional=list(args) (line 262)
-# ---------------------------------------------------------------------------
-
-
 @pytest.mark.asyncio
 async def test_call_modbus_no_signature_positional():
     """When _get_signature returns None, args are passed as positional (line 262)."""
@@ -528,46 +306,6 @@ async def test_call_modbus_no_signature_positional():
 
 # ---------------------------------------------------------------------------
 # get_rtu_framer coverage (lines 82-89)
-# ---------------------------------------------------------------------------
-
-
-def test_get_rtu_framer_returns_framer_type_rtu(monkeypatch):
-    """get_rtu_framer() returns FramerType.RTU when FramerType is available (lines 82-84)."""
-    from custom_components.thessla_green_modbus import modbus_helpers
-
-    class FakeFramerType:
-        RTU = "RTU_FRAMER"
-
-    monkeypatch.setattr(modbus_helpers, "FramerType", FakeFramerType)
-    result = modbus_helpers.get_rtu_framer()
-    assert result == "RTU_FRAMER"  # nosec B101
-
-
-def test_get_rtu_framer_returns_modbus_rtu_framer(monkeypatch):
-    """get_rtu_framer() returns ModbusRtuFramer when FramerType is None (lines 87-88)."""
-    from custom_components.thessla_green_modbus import modbus_helpers
-
-    class FakeRtuFramer:
-        pass
-
-    monkeypatch.setattr(modbus_helpers, "FramerType", None)
-    monkeypatch.setattr(modbus_helpers, "ModbusRtuFramer", FakeRtuFramer)
-    result = modbus_helpers.get_rtu_framer()
-    assert result is FakeRtuFramer  # nosec B101
-
-
-def test_get_rtu_framer_returns_none_when_both_unavailable(monkeypatch):
-    """get_rtu_framer() returns None when both FramerType and ModbusRtuFramer are None (line 89)."""
-    from custom_components.thessla_green_modbus import modbus_helpers
-
-    monkeypatch.setattr(modbus_helpers, "FramerType", None)
-    monkeypatch.setattr(modbus_helpers, "ModbusRtuFramer", None)
-    result = modbus_helpers.get_rtu_framer()
-    assert result is None  # nosec B101
-
-
-# ---------------------------------------------------------------------------
-# _call_modbus StopIteration path (lines 253-255)
 # ---------------------------------------------------------------------------
